Rename module pytensor.tensor.var to pytensor.tensor.variable
