+++ pytensor/tensor/sharedvar.py
@@ -6,7 +6,7 @@ from pytensor.compile import SharedVariable, shared_constructor
 from pytensor.misc.safe_asarray import _asarray
 from pytensor.tensor import _get_vector_length
 from pytensor.tensor.type import TensorType
-from pytensor.tensor.var import _tensor_py_operators
+from pytensor.tensor.variable import _tensor_py_operators
 
 
 def load_shared_variable(val):
